multimodal_deepagent/pyproject.toml
[project]
name = "multimodal-deepagent-a2a"
version = "0.1.0"
description = "Multimodal A2A agent that processes images/videos via MinIO + MCP tools"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "arize-phoenix>=13.19.2",
    "arize-phoenix-client>=2.1.0",
    "arize-phoenix-otel>=0.15.0",
    "click>=8.1.8",
    "deepagents>=0.4.12",
    "httpx>=0.28.1",
    "langchain>=1.2.13",
    "langchain-openai>=1.1.11",
    "langchain-mcp-adapters>=0.2.2",
    "langgraph>=1.1.3",
    "pydantic>=2.10.6",
    "python-dotenv>=1.1.0",
    "uvicorn>=0.34.2",
    "sse-starlette>=2.3.6",
    "starlette>=0.46.2",
    "a2a-sdk>=0.3.0,<1.0.0",
    "boto3>=1.35.0",
    "opencv-python-headless>=4.11.0",
    "pdfminer.six>=20221105",
    "python-docx>=1.1.0",
    "openinference-instrumentation-langchain>=0.1.61",
    "watchdog>=6.0.0",
]
